[refactor] remove circle imports
